Improve tls tests
