+++ rust/crd/src/lib.rs
@@ -905,66 +905,4 @@ mod tests {
             Some("testcluster-router.default.svc.cluster.local".to_string())
         )
     }
-
-    #[test]
-    fn test_druid_cluster_config_tls() {
-        let input = r#"
-        deepStorage:
-          hdfs:
-            configMapName: druid-hdfs
-            directory: /druid
-        metadataStorageDatabase:
-          dbType: derby
-          connString: jdbc:derby://localhost:1527/var/druid/metadata.db;create=true
-          host: localhost
-          port: 1527
-        zookeeperConfigMapName: zk-config-map
-        "#;
-        let druid_cluster_config: DruidClusterConfig =
-            serde_yaml::from_str(input).expect("illegal test input");
-
-        assert_eq!(
-            druid_cluster_config.zookeeper_config_map_name,
-            "zk-config-map".to_string()
-        );
-        assert_eq!(
-            druid_cluster_config.tls,
-            Some(DruidTls {
-                server_and_internal_secret_class: Some("tls".to_string())
-            }),
-        );
-        assert_eq!(druid_cluster_config.authentication, vec![],);
-
-        let input = r#"
-        authentication:
-          - authenticationClass: druid-user-authentication-class
-        tls:
-          serverAndInternalSecretClass: druid-secret-class
-        zookeeperConfigMapName: zk-config-map
-        metadataStorageDatabase:
-          dbType: derby
-          connString: jdbc:derby://localhost:1527/var/druid/metadata.db;create=true
-          host: localhost
-          port: 1527
-        deepStorage:
-          hdfs:
-            configMapName: druid-hdfs
-            directory: /druid
-        "#;
-        let druid_cluster_config: DruidClusterConfig =
-            serde_yaml::from_str(input).expect("illegal test input");
-
-        assert_eq!(
-            druid_cluster_config.tls,
-            Some(DruidTls {
-                server_and_internal_secret_class: Some("druid-secret-class".to_string())
-            }),
-        );
-        assert_eq!(
-            druid_cluster_config.authentication,
-            vec![DruidAuthentication {
-                authentication_class: "druid-user-authentication-class".to_string()
-            }],
-        );
-    }
 }
